Create useAddUserToRoom custom hook
